Implemented rudimentary progress updates for FTP.
The implementation is fairly straightforward:

1. Create a list of directories of be created and a list of files to
uploaded.
2. Add up the total number of "operations" between those two lists.
3. Run through all the items in each list, and after performing each
operation, increment a counter. Calculate progress as (completed
operations)/(total operations).

This obviously doesn't take into account varying file sizes, and the
fact that creating a directory is much less intensive than uploading a
file. But at least now there's a working progress bar!

+++ Mac App/Controllers/TBPublishSheetController.m
@@ -36,7 +36,7 @@
 	
 	[self.publisher setProgressHandler:^(NSInteger progress, NSInteger total) {
 		self.indicator.indeterminate = NO;
-		self.indicator.doubleValue = (progress/total) * 100;
+		self.indicator.doubleValue = ((double)progress/(double)total) * 100.0;
 	}];
 	
 	[self.publisher setCompletionHandler:^() {
